more gui classes extracted
