+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/Browser.java
@@ -79,8 +79,11 @@ import com.swirlds.config.api.Configuration;
 import com.swirlds.config.api.ConfigurationBuilder;
 import com.swirlds.config.api.source.ConfigSource;
 import com.swirlds.fchashmap.config.FCHashMapConfig;
-import com.swirlds.gui.GuiModel;
 import com.swirlds.gui.WindowConfig;
+import com.swirlds.gui.model.GuiModel;
+import com.swirlds.gui.model.InfoApp;
+import com.swirlds.gui.model.InfoMember;
+import com.swirlds.gui.model.InfoSwirld;
 import com.swirlds.jasperdb.config.JasperDbConfig;
 import com.swirlds.logging.payloads.NodeAddressMismatchPayload;
 import com.swirlds.logging.payloads.NodeStartPayload;
@@ -99,9 +102,6 @@ import com.swirlds.platform.event.preconsensus.PreconsensusEventStreamConfig;
 import com.swirlds.platform.event.tipset.EventCreationConfig;
 import com.swirlds.platform.gossip.chatter.config.ChatterConfig;
 import com.swirlds.platform.gossip.sync.config.SyncConfig;
-import com.swirlds.platform.gui.internal.InfoApp;
-import com.swirlds.platform.gui.internal.InfoMember;
-import com.swirlds.platform.gui.internal.InfoSwirld;
 import com.swirlds.platform.gui.internal.StateHierarchy;
 import com.swirlds.platform.health.OSHealthChecker;
 import com.swirlds.platform.health.clock.OSClockSpeedSourceChecker;
