+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/internal/SwirldMenu.java
@@ -22,8 +22,8 @@ import static com.swirlds.logging.LogMarker.EXCEPTION;
 import com.swirlds.common.system.Platform;
 import com.swirlds.common.system.SystemExitCode;
 import com.swirlds.common.utility.PlatformVersion;
-import com.swirlds.gui.GuiModel;
 import com.swirlds.gui.components.ScrollableJPanel;
+import com.swirlds.gui.model.GuiModel;
 import edu.umd.cs.findbugs.annotations.Nullable;
 import java.awt.BorderLayout;
 import java.awt.Color;
