+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/internal/WinTab2Consensus.java
@@ -18,11 +18,11 @@ package com.swirlds.platform.gui.internal;
 
 import static com.swirlds.gui.GuiUtils.wrap;
 
-import com.swirlds.gui.GuiModel;
+import com.swirlds.common.system.Platform;
 import com.swirlds.gui.GuiUtils;
 import com.swirlds.gui.components.PrePaintableJPanel;
+import com.swirlds.gui.model.GuiModel;
 import com.swirlds.platform.Consensus;
-import com.swirlds.platform.SwirldsPlatform;
 import com.swirlds.platform.components.state.StateManagementComponent;
 import com.swirlds.platform.gui.GuiPlatformAccessor;
 import com.swirlds.platform.state.signed.SignedStateInfo;
@@ -56,7 +56,7 @@ class WinTab2Consensus extends PrePaintableJPanel {
             if (WinBrowser.memberDisplayed == null) {
                 return;
             }
-            SwirldsPlatform platform = WinBrowser.memberDisplayed.getPlatform();
+            Platform platform = WinBrowser.memberDisplayed.getPlatform();
             String s = "";
             s += GuiModel.getInstance().getPlatformName(platform.getSelfId());
             final Consensus consensus = GuiPlatformAccessor.getInstance().getConsensus(platform.getSelfId());
