+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/gui/internal/WinTabSwirlds.java
@@ -21,8 +21,11 @@ import static com.swirlds.logging.LogMarker.EXCEPTION;
 import static com.swirlds.platform.gui.internal.BrowserWindowManager.getStateHierarchy;
 
 import com.swirlds.gui.GuiConstants;
-import com.swirlds.gui.InfoEntity;
 import com.swirlds.gui.components.PrePaintableJPanel;
+import com.swirlds.gui.model.InfoApp;
+import com.swirlds.gui.model.InfoEntity;
+import com.swirlds.gui.model.InfoMember;
+import com.swirlds.gui.model.InfoSwirld;
 import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.FlowLayout;
