+++ platform-sdk/swirlds-platform-gui/src/main/java/com/swirlds/gui/model/InfoApp.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2018-2023 Hedera Hashgraph, LLC
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package com.swirlds.platform.gui.internal;
+package com.swirlds.gui.model;
 
-import com.swirlds.gui.InfoEntity;
 import java.util.ArrayList;
 import java.util.List;
