+++ platform-sdk/swirlds-platform-gui/src/main/java/com/swirlds/gui/model/InfoEntity.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.swirlds.gui;
+package com.swirlds.gui.model;
 
 import javax.swing.JPanel;
