+++ platform-sdk/swirlds-platform-gui/src/main/java/com/swirlds/gui/model/WordList.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016-2023 Hedera Hashgraph, LLC
+ * Copyright (C) 2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.swirlds.platform.gui.internal;
+package com.swirlds.gui.model;
 
 import java.util.Arrays;
 import java.util.List;
@@ -26,9 +26,9 @@ import java.util.List;
  *
  * These are 2048 English words, no two of which have the same first 4 letters.
  */
-class WordList {
+public class WordList {
     /** the list of 2048 words */
-    static List<String> words = Arrays.asList(
+    public static List<String> words = Arrays.asList(
             "abandon",
             "ability",
             "able",
